requirements.txt
Pillow==9.5.0
matplotlib==3.8.1
opencv-python==4.8.1.78